fix(build): clear -Werror tail under clang-tidy and gcc-14
The dispatch and chain/scan headers fail -Werror under clang-tidy;
the failures were shadowed by GoogleTest's own tidy errors and
surface once those are scoped out. Two need scoped NOLINTs:
`unique_ptr<std::atomic<T>[]>` and a templated-noexcept dispatcher
whose throw/no-throw branch tidy cannot see through. The opt-in
padding analyzer is disabled globally because cache-line alignment
is a deliberate false-sharing trade-off and the analyzer fires per
instantiation, defeating per-decl NOLINTs. gcc-14 separately flags
`-Wsign-conversion` at `CPU_SET(cpu, ...)`, fixed with a `size_t`
cast.

+++ include/citor/detail/lookback_scan.h
@@ -34,7 +34,7 @@ template <class T>
 struct alignas(kCacheLine) LookbackTile {
   /// State-machine value stored in `flag`. The transitions are monotonic:
   /// `Initialized` -> `AggregateAvailable` -> `PrefixAvailable`.
-  enum class Flag : std::uint64_t {
+  enum class Flag : std::uint8_t {
     /// The tile owns a slot but has not yet computed its local aggregate.
     Initialized = 0,
     /// `aggregate` is published and synchronises through an acquire-load
